decouple pcap part
